fix no-default-features build, move tx JSON related code into json module;
